Исправлены методы чтения данных с консоли. Исправлен тип переменной r…

+++ src/by/tms/domain/Admin.java
@@ -2,11 +2,11 @@ package by.tms.domain;
 
 public class Admin extends UsersOfStudentBase {
 
-	public Admin(long id, String name, String login, String password, long role) {
+	public Admin(long id, String name, String login, String password, int role) {
 		super(id, name, login, password, role);
 	}
 
-	public Admin(String name, String login, String password, long role) {
+	public Admin(String name, String login, String password, int role) {
 		super(name, login, password, role);
 	}
